+++ src/lib/isISO31661Alpha2.js
@@ -1,4 +1,5 @@
 import assertString from './util/assertString';
+import includes from './util/includes';
 
 // from https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2
 const validISO31661Alpha2CountriesCodes = [
@@ -31,5 +32,5 @@ const validISO31661Alpha2CountriesCodes = [
 
 export default function isISO31661Alpha2(str) {
   assertString(str);
-  return validISO31661Alpha2CountriesCodes.includes(str.toUpperCase());
+  return includes(validISO31661Alpha2CountriesCodes, str.toUpperCase());
 }