+++ src/lib/isISO31661Alpha3.js
@@ -1,4 +1,5 @@
 import assertString from './util/assertString';
+import includes from './util/includes';
 
 // from https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3
 const validISO31661Alpha3CountriesCodes = [
@@ -22,5 +23,5 @@ const validISO31661Alpha3CountriesCodes = [
 
 export default function isISO31661Alpha3(str) {
   assertString(str);
-  return validISO31661Alpha3CountriesCodes.includes(str.toUpperCase());
+  return includes(validISO31661Alpha3CountriesCodes, str.toUpperCase());
 }